[eas-cli] Consolidate supabase helpers; move writeEnvVarsAsync
Drop constant-only sdk.ts/environments.ts into env.ts. Move the
generic writeEnvVarsAsync loop into environments/variables.
