--- packages/eas-cli/src/integrations/supabase/environments.ts
@@ -1,8 +1,0 @@
-import { DefaultEnvironment } from '../../build/utils/environment';
-
-// Production-first, matching the PostHog and Convex integrations.
-export const EAS_SUPABASE_ENVIRONMENTS = [
-  DefaultEnvironment.Production,
-  DefaultEnvironment.Preview,
-  DefaultEnvironment.Development,
-];
