--- packages/eas-cli/src/integrations/supabase/sdk.ts
@@ -1,5 +1,0 @@
-/** Packages installed by `eas integrations:supabase:connect`. */
-export const SDK_PACKAGES = ['@supabase/supabase-js', 'react-native-url-polyfill', 'expo-sqlite'];
-
-/** Config plugin added alongside the SDK packages. */
-export const CONFIG_PLUGIN = 'expo-sqlite';
